backend/requirements.txt
djangorestframework
mysqlclient == 2.1.1
